[integ-tests] reorganize p4d tests
* Merge nccl test in test_hit_efa
* Rename test_hit_efa to test_efa
* Run test_efa on all OSes with p4d instances
* Reduce number of p4d instances required by test_efa test
* Move all p4d tests to us-west-2
