+++ tests/integration-tests/configs/p4d.yaml
@@ -5,7 +5,7 @@
 ---
 test-suites:
   efa:
-    test_efa.py::test_hit_efa:
+    test_efa.py::test_efa:
       dimensions:
         - regions: {{ regions }}
           instances: {{ instances }}
