+++ tests/integration-tests/tests/efa/test_efa/test_efa/init_osu_benchmarks.sh
