+++ tests/integration-tests/tests/efa/test_efa/test_efa/osu_benchmarks/config.sub
