+++ tests/integration-tests/tests/efa/test_efa/test_efa/osu_benchmarks/results/c5n.18xlarge/intelmpi/osu_allgather
