+++ tests/integration-tests/tests/efa/test_efa/test_efa/osu_collective_submit_openmpi.sh
