+++ tests/integration-tests/tests/efa/test_efa/test_efa/osu_pt2pt_submit_intelmpi.sh
