--- tests/integration-tests/tests/nccl/test_nccl/test_nccl/pcluster.config.yaml
@@ -1,32 +1,0 @@
-Image:
-  Os: {{ os }}
-HeadNode:
-  InstanceType: {{ instance }}
-  Networking:
-    SubnetId: {{ public_subnet_id }}
-    ElasticIp: true
-  Ssh:
-    KeyName: {{ key_name }}
-  Imds:
-    Secured: {{ imds_secured }}
-Scheduling:
-  Scheduler: {{ scheduler }}
-  SlurmQueues:
-    - Name: ondemand1
-      ComputeResources:
-        - Name: ondemand-i1
-          InstanceType: {{ instance }}
-          DisableSimultaneousMultithreading: true
-          MinCount: 2
-          Efa:
-            Enabled: true
-            GdrSupport: true
-      Networking:
-        PlacementGroup:
-          Enabled: true
-        SubnetIds:
-          - {{ private_subnet_id }}
-SharedStorage:
-  - MountDir: /shared
-    Name: name1
-    StorageType: Ebs
